internal/providers/ec2: fix ssh keys
A recent commit resulted in the ec2 fetching logic to need to fetch a
resource from the dynamic endpoint in the ec2 metadata service in
addition to the meta-data endpoint. This resulted in all calls to fetch
a resource now needing to specify an additional part of the url prefix.
The logic for fetching ssh keys was overlooked when this change was
made.

This commit adds this meta-data prefix to the ssh key fetching logic.