providers: added azure ssh-keys support
this is the first provider to be implemented. this adds all of the
dependencies and the functionality needed to decrypt the ssh keys given
to us by the azure metadata service. the code should be well commented
enough to explain exactly what has to be done to actually get the
information (it's not straightforward).