providers/azure: wait until dhcp lease is aquired
There was an issue with the golang version where it would sometimes run
before the dhcp lease was actually aquired, so it would fail to get the
ip address (coreos/coreos-metadata#26). Let's not make the same mistake.

I would prefer it if this was a backoff timer of some kind instead of
just looping forever, but for now, we are just going to do what the
golang implementation does.