main: accept long args with one hyphen
the "go style" flags where flags can be accepted with either one or two
hyphens at the beginning needs to be supported, since we are trying to
match the original go interface as closely as possible. the solution is
to do a little preprocessing on the arguments that are passed, where if
a flag with a single hyphen is passed, we add another hyphen. this only
works because the flags we have defined don't have short arguments. we
would have to be more careful if short arguments are ever added.

fixes coreos/bugs#2240