providers/oracle: use default when key is missing
the oracle metadata service doesn't require that a machine has an ssh
key assigned to it. when an ssh key is missing, it returns metadata
without the `ssh_authorized_keys` field in the json, which caused serde
to fail when attempting to deserialize it. this change makes serde use
the default instance for string (which is "") if it can't find the field
when deserializing.