providers/packet: respect bond interface requests
the metadata for network interfaces has a bond field. up until now, we
ignored it and just always created a bond named bond0, attached all
addresses to it, and bonded both physical interfaces to it. this worked
because that's how the network was setup on most packet instances, but
it breaks because on instance type 1E there is an interface that
shouldn't be a part of the bond.

now, we create any bond interfaces mentioned in the network interfaces
list, so there may be multiple bonds. it also doesn't configure bonds
for interfaces that don't have a bond key listed in the metadata, which
fixes the behavior on packet instance type 1E.