network: add unmanaged field to Interface
on packet, most of their machines have multiple nics. by default on most
instance types, these nics are bonded together. on some instance types,
particularly 1E, this is not the case, and instead, some interfaces are
not configured to use any bond devices. when this happens, the interface
should be configured to be Unmanaged. this commit adds the ability to
mark some interfaces as unmanaged. it also updates the packet provider
to configure an interface to be unmanaged if there are no bond devices
configured for it.