travis: add a clippy pass
This moves clippy setup to an automatic CI check enforced via travis.
The side-effect of this is that we will be now automatically checking
every PR against clippy.

Travis pass is pinned to a specific clippy+nightly pair, which will
need to be bumped from time to time in the future to opt-in new lints.