packet: use correct phone home URL
First, we were incorrectly building the `/events` URL. But second,
POSTing a "succeeded" message is not a substitute for phoning home; we
still need to do that. In which case, there's no need to POST to
`/events` also. So let's just simplify things and do a POST only to
`/phone-home`, just like `packet-phone-home.service`.