provider/packet: add gateway to attributes
The Packet metadata service provides full information regarding a
machine's network configuration. This information includes the
gateway address assigned to a network interface (which is the Packet
ToR switch to which the machine is connected).

Having the gateway address in the attributes file is useful because
it allows systemd units to read this information via an environment
file. Without this information, units have to run a shell command
to calculate the gateway address, which can lead to "hacky" units and
is also error-prone.

In the match arms, `a` is renamed to `ip` so that `a` doesn't shadow
the `a` from the for loop.