providers/vmware: support injecting custom network kargs
This implements fetching logic specific to the `vmware` provider
in order to support injecting custom network kargs, which are meant
to override the default ones.

In particular, this checks for an optional guestinfo property named
`guestinfo.afterburn.initrd.network-kargs` which can be used to provide
a string containing VM-specific overrides.