azure: rework ready-state posting
This refactors the Azure provider, in order to better report instance
ready-state to the WireServer.
In particular, it looks like the platform expects a matching
`GoalStateIncarnation` to be posted in ready-state requests.
For this reason, this stops caching the goalstate content when creating
the provider. Instead, the goalstate is now fetched just in time,
and the incarnation number is used when posting the ready-state.
This also ensures that on transient errors all the endpoints are retried
long enough.