providers: add qemu provider for boot check-in
This adds a new provider for the "qemu" platform.
In particular, this provider currently supports checking in to
the host in order to report a successful boot.
This is performed through a VirtIO port, re-using the same console
and protocol used by oVirt guest-agent. Though, this is intentionally
not aimed at providing a full guest-agent implementation.