providers/vmware: allow avoiding iopl permission errors
This introduces a fallback path for the backdoor-opening logic,
in order to allow consuming applications to keep going when changing
I/O access level is not allowed.
Privileged mode is in theory more reliable but `kernel_lockdown(7)`
may block it due to `iopl()` usage. Thus, we try that first and
fall back if kernel blocks it.