cargo: remove carets from semver requirements
This removes all carets (`^`) from dependencies, resulting in
semver entries which are semantically equivalent from cargo
point of view.
This repository is starting to accumulate stale dependencies, as
dependabot is having troubles when trying to update them. I think
I'm seeing a pattern where the `^` may be the one thing confusing
the bot, and the changes in here try to verify such theory.

Ref: https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#specifying-dependencies-from-cratesio