providers: add Azure Stack Hub (azurestack)
Adds support and tests for the Azure Stack platform. Sets the hostname
based on the vm name. Obtains SSH keys. Checks in to finish the boot
process. Unlike Azure, does not offer any attributes. Follows the
same pattern as Azure with some notable differences:

The hostname is obtained by parsing the vm name from the identity
metadata service, similar to the WALinuxAgent. Early attempts were
made to obtain this vm name from the compute instance metadata service
which was available in a pre-release lab environment, but the endpoint
was not reachable pre-check in. Consdering the vm name was available
from the identity endpoint, accessing the instance metadata service
can be saved for future work.

SSH key functionality is the same as Azure.

Regarding attributes, obtaining the IP addresses in the same manner as
Azure led to incorrect values, so they were removed as well as vmsize
for simplicity. Therefore no attributes are available.

Check in is the same process as Azure.

Co-authored-by: Ben Howard <ben.howard@redhat.com>
Co-authored-by: Jonathan Lebon <jonathan@jlebon.com>