Add PowerVS provider
As part of supporting Power Virtual server as a platform (https://github.com/coreos/fedora-coreos-tracker/issues/817), Afterburn will be used to set the hostname attributes and possibly the ssh-keys.

This provider is very similar to the IBMCloud classic provider with the exception of the filesystem for the config drive which is `iso9660` similar to Openstack. The contents of the config drive are usually the meta_data.json and the network_data.json. This provider includes support for:
- hostname
- ssh keys