Explicitly log the hostname we write
This would have greatly helped debugging https://bugzilla.redhat.com/show_bug.cgi?id=2008521#c40
because it was not immediately obvious from the logs that
afterburn is just writing a full, untruncated hostname to `/sysroot/etc/hostname`.

We see in the logs that the service started/stopped successfully,
but in this case it happens that systemd truncates it, which obscured
the value.