build(deps): bump regex from 1.5.4 to 1.5.5
As this is only used via mockito, I assume that afterburn is not
vulnerable as it doesn't take a user supplied regex. This fixes a denial
of service security bug in the regex crate, just to be sure. The bug
made it possible for user supplied regexes to avoid mitigations which
are supposed to prevent the regex from taking a very long time to parse
with.

Fixes: https://rustsec.org/advisories/RUSTSEC-2022-0013
Fixes: CVE-2022-24713