cargo: update version ranges for post-1.x deps
The idea of the explicit ranges is to exactly replicate the ranges that
Cargo would allow if a bare version were specified, while convincing
Dependabot not to bump the minimum version.  For major versions > 0,
Cargo would allow every minor up to the next major, so we should do that
too.

See discussion here:
https://src.fedoraproject.org/rpms/rust-coreos-installer/pull-request/40