util: support DHCP option lookup from NetworkManager
Check both NetworkManager and networkd on each retry iteration.  The
NetworkManager check works with both dhclient and NM's internal DHCP
client.  Avoids a 30-second boot delay on Azure and AzureStack as we time
out waiting for a networkd lease.  Doesn't help in the initrd on RHCOS 8,
since NetworkManager doesn't daemonize there.

Support zbus 2.3 because Fedora isn't shipping zbus 3 yet.  The two are
API-compatible for our purposes.

Tested on Azure on Flatcar and RHEL in the real root, and FCOS in the
initrd.  Not tested on CloudStack, except to verify that NetworkManager
exposes a dhcp_server_identifier DHCP option with the expected format.

Fixes https://github.com/coreos/afterburn/issues/146.