providers/vmware: Process guestinfo.metadata netplan configuration
The network environment can be dynamic and thus needs to be provided as
VM metadata. Since the format should not depend on whether the VM runs
uses Ignition and Afterburn or Cloud-Init, the idea is to also support
the guestinfo.metadata variable as used by Cloud-Init which contains
Netplan YAML/JSON network configuration.
Add a new command to write out Netplan configs to a given directory,
similar as we do with networkd units. (While this is currently just used
for VMware, other providers could also construct the netplan data type
from the netplan-types crate to provide Netplan configurations if the OS
rather wants to use NetworkManager than systemd-networkd. For backwards
compatibility and to not need Netplan it would be nice to keep the
systemd-networkd support as long as its used.)

References:
https://cloudinit.readthedocs.io/en/latest/reference/datasources/vmware.html#walkthrough-of-guestinfo-keys-transport
https://cloudinit.readthedocs.io/en/latest/reference/network-config-format-v2.html
https://netplan.io/reference/
https://linux-on-z.blogspot.com/p/using-netplan-on-ibm-z.html