docs/release-notes: note Azure SSH regression fix with new openssl
On versions of the openssl crate new enough to have parse2(), the legacy
parse() function is a wrapper that panics when parsing the Azure SSH key
certificate:

    thread 'main' panicked at 'called `Option::unwrap()` on a `None` value', [...]/openssl-0.10.52/src/pkcs12.rs:42:31

We've switched to parse2() in the current codebase, but older versions
built with newer openssl will fail.  Note the fix.

See https://github.com/coreos/fedora-coreos-tracker/issues/1492.