openstack: Add attribute OPENSTACK_INSTANCE_UUID
In the OpenStack attributes, the field `OPENSTACK_INSTANCE_ID` reports
the instance ID fetched from the EC2-compatibility-layer metadata.
However, that identifier (in the format `i-something`) does not match
the one used to refer to machines in the OpenStack API.

With this patch, I propose to add a new OpenStack attribute
`OPENSTACK_INSTANCE_UUID` that reports the actual identifier for
OpenStack instances in the context of the OpenStack API.

Implementation-wise, this patch adds enables fetching the UUID from both
the config-drive and from the Nova metadata service endpoint. In order
to retain compatibility with old versions of the metadata service, this
patch uses the old endpoint `2012-08-10`.

Fixes #930