workflows/rust: directly update toolchain to 1.75.0
I have made a MR into repo-templates to update this file.
seen here: https://github.com/coreos/repo-templates/pull/230
Once it lands the downstream MR for updating the toolchain will
not be able to merge due to failing CI we currently have, and
we cannot merge the fix for the failing CI without it. So
add the lint CI update here to allow CI fixes to be mergable.